pyproject.toml
[tool.poetry]
name = "contacts-api"
version = "0.1.0"
description = "REST API for contact management"
authors = ["Your Name <your.email@example.com>"]
packages = [
    { include = "app" }
]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
sqlalchemy = "^2.0.22"
psycopg2-binary = "^2.9.9"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
python-dotenv = "^1.0.0"
alembic = "^1.12.0"
email-validator = "^2.1.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api" 